changed parsing to double instead of int

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -4,7 +4,7 @@ public class StringUtils {
     public static boolean isPositiveNumber(String str) {
         //code!
         try {
-            return Integer.parseInt(str) > 0;
+            return Double.parseDouble(str) > 0;
         } catch (NumberFormatException e) {
             return false;
         }
